deployments/nsfw-detector/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
python-multipart>=0.0.6
httpx>=0.25.0

